fixed：Issue #22：MyBatis组件的remove相关方法必须传入参数。

+++ acooly-components/acooly-component-mybatis/src/main/java/com/acooly/module/mybatis/ex/RemoveMapper.java
@@ -11,6 +11,9 @@ package com.acooly.module.mybatis.ex;
 
 import org.apache.ibatis.annotations.UpdateProvider;
 import org.apache.ibatis.mapping.MappedStatement;
+import org.springframework.dao.DataAccessException;
+import org.springframework.dao.UncategorizedDataAccessException;
+import org.springframework.jdbc.SQLWarningException;
 import tk.mybatis.mapper.mapperhelper.MapperHelper;
 import tk.mybatis.mapper.mapperhelper.SqlHelper;
 import tk.mybatis.mapper.provider.base.BaseDeleteProvider;
@@ -48,10 +51,16 @@ public interface RemoveMapper<T> {
         }
 
         public String removeById(MappedStatement ms) {
+            if (ms.getConfiguration().getVariables() == null) {
+                throw new SQLWarningException("删除时必须加参数",null);
+            }
             return super.deleteByPrimaryKey(ms);
         }
 
         public String remove(MappedStatement ms) {
+            if (ms.getConfiguration().getVariables() == null) {
+                throw new SQLWarningException("删除时必须加参数",null);
+            }
             return super.delete(ms);
         }
